fixed expressions between arrays with more than one dimension and axes which are not the same objects
np.array_equal([selfdim1, selfdim2], [otherdim1, otherdim2]) is wrong: it cannot be converted to a "pure" ndarray if the dimensions do not have the same size.

It also "worked" if selfdim1 and otherdim1 were in fact the same object, which they actually were in simple expressions. But this was not the intention of the test.