experiment with using numba to speed up link functions
for large datasets, we get an almost 2x speedup for all of them and 10x for min/max

Ultimately, I don't know whether we should rather do id2rownum[source_ids] outside the func and potentially cache the result, or do it inside the func (faster than non-caching).
Since we do not/cannot cache that kind of stuff yet, I guess I should go for the "inline" version for now but I will have to add a comment to keep that in mind.