experiment on using numba to speed up gini()
This is only a moderate success (~15% speed increase) as most of the time is spent in sorting the data and there is little we can do about that.
Using the O(n²) version of the gini definition (which does not require sorting) is already slower with n=1000, so not worth it.